FIX: compiler warning
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@2672 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Base/cmtkGaussianKernel.h
@@ -74,7 +74,7 @@ public:
 					    const TFloat maxError = 1e-5 /*!< Maximum approximation error: the kernel radius is computed so that truncated elements are below this value */ )
   {
     const double normFactor = 1.0/(sqrt(2*M_PI) * sigma.Value());
-    const size_t radius = Self::GetRadius( sigma, normFactor, maxError );
+    const size_t radius = static_cast<size_t>( Self::GetRadius( sigma, normFactor, maxError ) );
     
     std::vector<TFloat> kernel( radius + 1 );
     for ( size_t i = 0; i <= radius; ++i )
